Split header context from `Layout` and update imports across components
